+++ core-api/src/main/java/blog/jungmini/me/application/UserService.java
@@ -4,11 +4,10 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import blog.jungmini.me.common.error.CustomException;
+import blog.jungmini.me.common.error.ErrorType;
 import blog.jungmini.me.database.entity.UserEntity;
 import blog.jungmini.me.database.repository.UserRepository;
-
-import blog.jungmini.common.error.CustomException;
-import blog.jungmini.common.error.ErrorType;
 
 @Service
 public class UserService {
