+++ core-support/common/src/main/java/blog/jungmini/me/common/error/ErrorCode.java
@@ -1,4 +1,4 @@
-package blog.jungmini.common.error;
+package blog.jungmini.me.common.error;
 
 public enum ErrorCode {
     E400,
