+++ core-support/common/src/main/java/blog/jungmini/me/common/error/ErrorMessage.java
@@ -1,4 +1,4 @@
-package blog.jungmini.common.error;
+package blog.jungmini.me.common.error;
 
 import lombok.Getter;
