+++ core-support/common/src/main/java/blog/jungmini/me/common/error/HttpStatus.java
@@ -1,4 +1,4 @@
-package blog.jungmini.common.error;
+package blog.jungmini.me.common.error;
 
 public enum HttpStatus {
     BAD_REQUEST,
